feat: move discussion page into discussion sub-feature folder
Doing this so renames are applied correctly for maintaining git history
