+++ app/frontend/src/AppRoutes.tsx
@@ -15,7 +15,7 @@ import { CompleteResetPassword, ResetPassword } from "./features/auth/password";
 import Settings from "./features/auth/Settings";
 import Signup from "./features/auth/signup/Signup";
 import CommunityPage from "./features/communities/CommunityPage";
-import DiscussionPage from "./features/communities/DiscussionPage";
+import { DiscussionPage } from "./features/communities/discussion";
 import GroupPage from "./features/communities/GroupPage";
 import NewGuidePage from "./features/communities/NewGuidePage";
 import NewPlacePage from "./features/communities/NewPlacePage";
