+++ app/frontend/src/features/communities/discussion/DiscussionPage.tsx
