+++ app/frontend/src/features/communities/discussion/index.ts
@@ -1,0 +1,1 @@
+export { default as DiscussionPage } from "./DiscussionPage";
